Express Validation

+++ controllers/budget/deleteBudget.js
@@ -1,8 +1,15 @@
+const { body, validationResult } = require('express-validator');
 const BudgetModel = require('../../models/budget');
 
 async function deleteBudget(req, res) {
-    const { budgetId } = req.body;
+    await body('budgetId').notEmpty().withMessage('Budget ID is required').isMongoId().withMessage('Invalid Budget ID').run(req);
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+    }
+
     try {
+        const { budgetId } = req.body;
         const deletedBudget = await BudgetModel.findOneAndDelete({ _id: budgetId });
         if (deletedBudget) {
             return res.status(204).json({ message: 'Budget deleted successfully' });
